+++ tests/amt_user_test.py
@@ -1,7 +1,7 @@
 """ test user functionality """
 
 import pytest_check as check
-import acct_mgt_test as amt
+import amt_helper as amt
 
 
 def test_user(acct_mgt_url, auth_opts):
